requirements.txt
tensorflow-cpu==2.13.0
mediapipe==0.10.9
opencv-python-headless==4.8.1.78
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.0.0
numpy==1.24.3
